chore: fix glommio features
